frontend: rendering: setup browser interface class
I encountered some problems with jest and js-environment-jsdom when trying to
test calls like `Element.getClientRects()`. There are ways and libraries to
mock these calls but they were pretty complicated to set up and didn't look
too reliable to me when it comes to library updates.

To avoid mocking, this library adds a new class called `BrowserInterface` which
abstracts browser calls that are not covered by js-environment-jsdom.

Signed-off-by: Florian Scherf <mail@florianscherf.de>