frontend: setup webpack
Previously, only first level imports in modules worked in jest tests. When a
module imported another module, using ES06 imports, the imports couldn't be
resolved.

This patch adds webpack to the build chain because it was the easiest way to
support imports that tsc and jest would understand.
An added benefit is that we have a smaller bundle size now.

Signed-off-by: Florian Scherf <mail@florianscherf.de>